Reorganize header files
Include required predeclarations and includes into header files so that
they can be either included individually as is done now, or so that the
`plorth/plorth.hpp` header can be the only header needed to include to
embed the interpreter into C++ application.
